+++ crates/julie-extractors/src/gdscript/helpers.rs
@@ -16,10 +16,10 @@ pub(super) fn extract_variable_annotations(
 
     // Check for annotations as children
     for i in 0..parent_node.child_count() {
-        if let Some(child) = parent_node.child(i) {
+        if let Some(child) = parent_node.child(i as u32) {
             if child.kind() == "annotations" {
                 for j in 0..child.child_count() {
-                    if let Some(annotation_child) = child.child(j) {
+                    if let Some(annotation_child) = child.child(j as u32) {
                         if annotation_child.kind() == "annotation" {
                             let annotation_text = base.get_node_text(&annotation_child);
                             annotations.push(annotation_text);
@@ -35,7 +35,7 @@ pub(super) fn extract_variable_annotations(
         // Find parent node index
         let mut node_index = None;
         for i in 0..grandparent.child_count() {
-            if let Some(child) = grandparent.child(i) {
+            if let Some(child) = grandparent.child(i as u32) {
                 if child.id() == parent_node.id() {
                     node_index = Some(i);
                     break;
@@ -48,10 +48,10 @@ pub(super) fn extract_variable_annotations(
 
             // Look backwards for annotations
             for i in (0..idx).rev() {
-                if let Some(child) = grandparent.child(i) {
+                if let Some(child) = grandparent.child(i as u32) {
                     if child.kind() == "annotations" {
                         for j in 0..child.child_count() {
-                            if let Some(annotation_child) = child.child(j) {
+                            if let Some(annotation_child) = child.child(j as u32) {
                                 if annotation_child.kind() == "annotation" {
                                     let annotation_text = base.get_node_text(&annotation_child);
                                     annotations.push(annotation_text.clone());
